Code looks better and is more organized. Created csv parser file that the retrieveSocrataApiData call. Can probably make a main file that calls the functions in both those scripts to retrieve all data.
